% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_c1_c2_c3.R
\name{as_dataframe}
\alias{as_dataframe}
\title{Convert simulation output to data frame}
\usage{
as_dataframe(sim)
}
\arguments{
\item{sim}{Output from rexp_series_md()}
}
\value{
Data frame with t, delta, x1, x2, ..., xm columns
}
\description{
Convert simulation output to data frame
}
