% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_c1_c3.R
\name{compute_pi_all}
\alias{compute_pi_all}
\title{Compute pi_k(c) for all k in c}
\usage{
compute_pi_all(c, P)
}
\arguments{
\item{c}{Logical vector indicating candidate set}

\item{P}{Inclusion probability matrix}
}
\value{
Named vector of pi_k values for k in c
}
\description{
Compute pi_k(c) for all k in c
}
