% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{dwei_series}
\alias{dwei_series}
\title{PDF of Weibull series system}
\usage{
dwei_series(t, shapes, scales, log = FALSE)
}
\arguments{
\item{t}{Time points (non-negative)}

\item{shapes}{Shape parameters (k_1, ..., k_m)}

\item{scales}{Scale parameters (beta_1, ..., beta_m)}

\item{log}{If TRUE, return log-density}
}
\value{
Density values
}
\description{
The system fails when the first component fails. The PDF is:
f(t) = h_sys(t) * R_sys(t) where h_sys = sum of component hazards
and R_sys = product of component reliabilities.
}
