% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{fit_c1_c2_c3}
\alias{fit_c1_c2_c3}
\title{Fit C1-C2-C3 model}
\usage{
fit_c1_c2_c3(sim)
}
\arguments{
\item{sim}{Simulated data}
}
\value{
Fitted model results
}
\description{
Fit C1-C2-C3 model
}
\keyword{internal}
