% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guo_weibull_series_md}
\alias{guo_weibull_series_md}
\title{Guo Weibull Series Masked Data}
\format{
A data frame with 30 rows and 4 columns:
\describe{
\item{t}{System failure time}
\item{x1,x2,x3}{Candidate set indicators}
}
}
\usage{
guo_weibull_series_md
}
\description{
Masked failure data from a 3-component Weibull series system,
based on the dataset from Guo et al. (2013).
}
\keyword{datasets}
