% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guo_weibull_series_mle}
\alias{guo_weibull_series_mle}
\title{Guo Weibull Series MLE Results}
\format{
A list with components:
\describe{
\item{mle}{MLE parameter estimates}
\item{loglike}{Log-likelihood at MLE}
\item{data}{The input data}
}
}
\usage{
guo_weibull_series_mle
}
\description{
Maximum likelihood estimation results for the Guo et al. (2013) dataset.
}
\keyword{datasets}
