% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{hazard_wei_series}
\alias{hazard_wei_series}
\title{Hazard function of Weibull series system}
\usage{
hazard_wei_series(t, shapes, scales)
}
\arguments{
\item{t}{Time points}

\item{shapes}{Shape parameters}

\item{scales}{Scale parameters}
}
\value{
Hazard values
}
\description{
Hazard function of Weibull series system
}
