% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_c1_c2_c3.R
\name{loglik_exp_series_df}
\alias{loglik_exp_series_df}
\title{Wrapper: Log-likelihood from data frame}
\usage{
loglik_exp_series_df(df, tvar = "t", deltavar = "delta", setvar = "x")
}
\arguments{
\item{df}{Data frame with columns t, delta, and x1, x2, ..., xm}

\item{tvar}{Column name for system lifetime (default "t")}

\item{deltavar}{Column name for censoring indicator (default "delta")}

\item{setvar}{Column prefix for candidate set (default "x")}
}
\value{
Log-likelihood function
}
\description{
Wrapper: Log-likelihood from data frame
}
