% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_c1_c2_c3.R
\name{loglik_wei_series}
\alias{loglik_wei_series}
\title{Log-likelihood for Weibull series system with masked data (C1, C2, C3)}
\usage{
loglik_wei_series(t, C, delta = NULL)
}
\arguments{
\item{t}{Numeric vector of system lifetimes}

\item{C}{Logical matrix of candidate sets (n x m)}

\item{delta}{Censoring indicators (1=observed, 0=censored)}
}
\value{
Function that takes theta = c(k1,b1,k2,b2,...) and returns log-likelihood
}
\description{
Returns a log-likelihood function for masked data from a Weibull series
system under conditions C1, C2, C3.
}
\details{
The log-likelihood is:
\verb{l(theta) = sum_i [-sum_j (t_i/beta_j)^k_j] + sum_\{i:delta_i=1\} log(sum_\{j in C_i\} (k_j/beta_j)(t_i/beta_j)^(k_j-1))}
}
