% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_c1_c2_c3.R
\name{loglik_wei_series_df}
\alias{loglik_wei_series_df}
\title{Wrapper: Log-likelihood (Weibull) from data frame}
\usage{
loglik_wei_series_df(df, tvar = "t", deltavar = "delta", setvar = "x")
}
\arguments{
\item{df}{Data frame with t, delta, x1, x2, ... columns}

\item{tvar, deltavar, setvar}{Column name parameters}
}
\value{
Log-likelihood function
}
\description{
Wrapper: Log-likelihood (Weibull) from data frame
}
