% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_c1_c2.R
\name{mle_exp_series_c1_c2_df}
\alias{mle_exp_series_c1_c2_df}
\title{Wrapper: MLE (relaxed C3) from data frame}
\usage{
mle_exp_series_c1_c2_df(
  df,
  alpha = NULL,
  base_p = 0.5,
  theta0 = NULL,
  tvar = "t",
  deltavar = "delta",
  setvar = "x"
)
}
\arguments{
\item{df}{Data frame with t, delta, x1, x2, ... columns}

\item{alpha}{Power parameter (NULL to estimate)}

\item{base_p}{Baseline inclusion probability}

\item{theta0}{Initial parameters}

\item{tvar, deltavar, setvar}{Column name parameters}
}
\value{
MLE result list
}
\description{
Wrapper: MLE (relaxed C3) from data frame
}
