% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_c1_c2_c3.R
\name{mle_wei_series_df}
\alias{mle_wei_series_df}
\title{Wrapper: MLE (Weibull) from data frame}
\usage{
mle_wei_series_df(
  df,
  tvar = "t",
  deltavar = "delta",
  setvar = "x",
  theta0 = NULL
)
}
\arguments{
\item{df}{Data frame with t, delta, x1, x2, ... columns}

\item{tvar, deltavar, setvar}{Column name parameters}

\item{theta0}{Initial parameters}
}
\value{
MLE result list
}
\description{
Wrapper: MLE (Weibull) from data frame
}
