% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_c1_c2_c3.R
\name{pack_wei_params}
\alias{pack_wei_params}
\title{Pack shapes and scales into Weibull parameter vector}
\usage{
pack_wei_params(shapes, scales)
}
\arguments{
\item{shapes}{Shape parameters}

\item{scales}{Scale parameters}
}
\value{
Parameter vector c(k1, beta1, k2, beta2, ..., km, betam)
}
\description{
Pack shapes and scales into Weibull parameter vector
}
\keyword{internal}
