% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{print_simulation_summary}
\alias{print_simulation_summary}
\title{Print simulation summary}
\usage{
print_simulation_summary(summary_df)
}
\arguments{
\item{summary_df}{Output from summarize_simulation}
}
\description{
Print simulation summary
}
