% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{pwei_series}
\alias{pwei_series}
\title{CDF of Weibull series system}
\usage{
pwei_series(q, shapes, scales, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{Quantile values}

\item{shapes}{Shape parameters}

\item{scales}{Scale parameters}

\item{lower.tail}{If TRUE (default), P(T <= q)}

\item{log.p}{If TRUE, return log probability}
}
\value{
Probabilities
}
\description{
CDF of Weibull series system
}
