% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{qwei_series}
\alias{qwei_series}
\title{Quantile function of Weibull series system}
\usage{
qwei_series(p, shapes, scales, lower.tail = TRUE, tol = 1e-10, max_iter = 100)
}
\arguments{
\item{p}{Probabilities}

\item{shapes}{Shape parameters}

\item{scales}{Scale parameters}

\item{lower.tail}{If TRUE (default), find q such that P(T <= q) = p}

\item{tol}{Convergence tolerance}

\item{max_iter}{Maximum iterations}
}
\value{
Quantile values
}
\description{
Uses Newton's method to invert the CDF.
}
