% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series_c1_c3.R
\name{rexp_series_md_c1_c3}
\alias{rexp_series_md_c1_c3}
\title{Generate masked data with general Bernoulli model (relaxed C2)}
\usage{
rexp_series_md_c1_c3(n, theta, P, tau, keep_latent = FALSE)
}
\arguments{
\item{n}{Sample size}

\item{theta}{Rate parameters for exponential components}

\item{P}{Inclusion probability matrix (m x m): \verb{P[j,k] = P(j in C | K = k)}}

\item{tau}{Right-censoring time}

\item{keep_latent}{If TRUE, include latent variables in output}
}
\value{
List with t, delta, C, k, and optionally Tm, P
}
\description{
Generate masked data with general Bernoulli model (relaxed C2)
}
