% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{run_all_simulations}
\alias{run_all_simulations}
\title{Run all simulation scenarios}
\usage{
run_all_simulations(config, scenarios = 1:4, ..., verbose = TRUE)
}
\arguments{
\item{config}{Simulation configuration}

\item{scenarios}{Which scenarios to run (default: all 1-5)}

\item{...}{Additional parameters}

\item{verbose}{Print progress}
}
\value{
Data frame with all results
}
\description{
Run all simulation scenarios
}
