% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{rwei_series}
\alias{rwei_series}
\title{Random generation from Weibull series system}
\usage{
rwei_series(n, shapes, scales)
}
\arguments{
\item{n}{Number of observations}

\item{shapes}{Shape parameters}

\item{scales}{Scale parameters}
}
\value{
Vector of system lifetimes
}
\description{
Random generation from Weibull series system
}
