% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_c1_c2_c3.R
\name{rwei_series_md}
\alias{rwei_series_md}
\title{Generate masked data from Weibull series system}
\usage{
rwei_series_md(n, shapes, scales, p, tau, keep_latent = FALSE)
}
\arguments{
\item{n}{Sample size}

\item{shapes}{Shape parameters (k_1, ..., k_m)}

\item{scales}{Scale parameters (beta_1, ..., beta_m)}

\item{p}{Probability each non-failed component is in candidate set}

\item{tau}{Right-censoring time}

\item{keep_latent}{If TRUE, include latent component times}
}
\value{
List with t, delta, C, k, m, shapes, scales
}
\description{
Generate masked data from Weibull series system
}
