% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_intervals.R
\name{simulate_violation}
\alias{simulate_violation}
\title{Generate one simulated dataset under a violation parameter.}
\usage{
simulate_violation(
  violation,
  alpha,
  n,
  shapes,
  scales,
  base_p,
  tau,
  C2_direction
)
}
\description{
Generate one simulated dataset under a violation parameter.
}
