% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{surv_wei_series}
\alias{surv_wei_series}
\title{Survival function of Weibull series system}
\usage{
surv_wei_series(t, shapes, scales, log.p = FALSE)
}
\arguments{
\item{t}{Time points}

\item{shapes}{Shape parameters}

\item{scales}{Scale parameters}

\item{log.p}{If TRUE, return log survival probability}
}
\value{
Survival probabilities
}
\description{
Survival function of Weibull series system
}
