% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_c1_c2_c3.R
\name{unpack_wei_params}
\alias{unpack_wei_params}
\title{Extract shapes and scales from Weibull parameter vector}
\usage{
unpack_wei_params(theta)
}
\arguments{
\item{theta}{Parameter vector c(k1, beta1, k2, beta2, ..., km, betam)}
}
\value{
List with shapes and scales vectors
}
\description{
Extract shapes and scales from Weibull parameter vector
}
\keyword{internal}
