% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_c1_c2.R
\name{wei_inclusion_probs}
\alias{wei_inclusion_probs}
\title{Compute normalized inclusion probabilities for Weibull power model}
\usage{
wei_inclusion_probs(shapes, scales, alpha, base_p)
}
\arguments{
\item{shapes}{Shape parameters}

\item{scales}{Scale parameters}

\item{alpha}{Power parameter}

\item{base_p}{Baseline inclusion probability}
}
\value{
Vector of inclusion probabilities (max = base_p)
}
\description{
Compute normalized inclusion probabilities for Weibull power model
}
\keyword{internal}
