% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{wei_series_mttf}
\alias{wei_series_mttf}
\title{Mean time to failure (MTTF) of Weibull series system}
\usage{
wei_series_mttf(shapes, scales, upper = NULL)
}
\arguments{
\item{shapes}{Shape parameters}

\item{scales}{Scale parameters}

\item{upper}{Upper integration limit (default: large value)}
}
\value{
MTTF
}
\description{
Computed by numerical integration of the survival function.
}
